[~ing] S03P31A409-102 : 취미 댓글 수정

+++ back/chimi-service/src/main/java/com/chimi/controller/CommentController.java
@@ -2,8 +2,10 @@ package com.chimi.controller;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
+import com.chimi.intercomm.UserClient;
 import com.chimi.model.ChimiComment;
 import com.chimi.model.CommentLike;
 import com.chimi.model.PKSet;
@@ -17,9 +19,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -36,35 +40,37 @@ public class CommentController {
     private CommentService commentService;
     @Autowired
     private LikeService likeService;
-
+    @Autowired
+    private UserClient userClient;
     @PostMapping
     @ApiOperation(value = "새 댓글 쓰기")
-    public ResponseEntity<String> insert(@RequestBody ChimiComment comment) {
-    
+    public ResponseEntity<String> insert(@RequestHeader("accessToken") String access,@RequestBody ChimiComment comment) {
+    	HashMap<String, Object> userinfo = userClient.getUserInfo(access);
+    	comment.setUserId((Long) userinfo.get("id"));
         ChimiComment newComment = commentService.save(comment);     // 댓글 저장
 
         if(newComment != null)	return new ResponseEntity<>("success", HttpStatus.OK);
 		else					return new ResponseEntity<>("fail", HttpStatus.BAD_REQUEST);
     }
 
-    @GetMapping
+    @GetMapping("/{hid}")
 	@ApiOperation(value = "해당 파티의 모든 댓글 조회")
-	public ResponseEntity<List<ChimiCommentList>> searchAll(@RequestParam(required = true) Long hid, @RequestParam(required = true) String email) throws Exception {
+	public ResponseEntity<List<ChimiCommentList>> searchAll(@PathVariable long hid) throws Exception {
        
         List<ChimiComment> comments = commentService.findByHid(hid);
         List<ChimiCommentList> commentList = new ArrayList<>();
         for(ChimiComment comment : comments){
-            commentList.add(new ChimiCommentList(comment, likeService.findById(new PKSet(email, comment.getCid())).isPresent()));
+            commentList.add(new ChimiCommentList(comment, likeService.findById(new PKSet(comment.getUserId(), comment.getCid())).isPresent()));
         }
         return new ResponseEntity<>(commentList, HttpStatus.OK);
     }
     
     @PutMapping
 	@ApiOperation(value = "댓글 수정 ")
-	public ResponseEntity<String> modifyComment(@RequestBody ChimiComment comment) {
-
+	public ResponseEntity<String> modifyComment(@RequestHeader("accessToken") String access,@RequestBody ChimiComment comment) {
+    	HashMap<String, Object> userinfo = userClient.getUserInfo(access);
         ChimiComment newComment = commentService.findById(comment.getCid()).get();
-        if(newComment != null){
+        if(newComment != null && newComment.getUserId()==((long) userinfo.get("id"))){
             newComment.setContent(comment.getContent());
             newComment = commentService.save(newComment);
         }
@@ -73,11 +79,11 @@ public class CommentController {
 	
 	}	
 
-    @DeleteMapping
+    @DeleteMapping("/{cid}")
 	@ApiOperation(value = "댓글 삭제")
-	public ResponseEntity<String> deleteComment(Long cid) {
-
-        if(commentService.findById(cid).isPresent()){
+	public ResponseEntity<String> deleteComment(@RequestHeader("accessToken") String access,@PathVariable Long cid) {
+    	HashMap<String, Object> userinfo = userClient.getUserInfo(access);
+        if(commentService.findById(cid).isPresent() && commentService.findById(cid).get().getUserId()==((long) userinfo.get("id"))){
             commentService.deleteById(cid);
 
             if(!likeService.findByLikePKId(cid).isEmpty()){
